third_party/blink/web_tests/platform/generic/http/tests/local/formdata/send-form-data-with-empty-file-filename-expected.txt

Test that filename passed to FormData.append() takes precedence over filename in File, even if empty.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Sending FormData containing one file with custom empty filename:
PASS filename is ''
PASS successfullyParsed is true

TEST COMPLETE
